+++ src/Math/GraphMetrics.cpp
@@ -1,5 +1,9 @@
 #include "Math/GraphMetrics.h"
+
+#include "Data/Graph.h"
 #include "Math/DjikstraAlg.h"
+
+#include "Util/ApplicationConfig.h"
 
 void GraphMetrics::computeGraphMetrics(Data::Graph * activeGraph)
 {
@@ -24,7 +28,8 @@ void GraphMetrics::computeGraphMetrics(Data::Graph * activeGraph)
 		i.next();
 
 		// compute node degree
-		float nodeDegree = ((float) i.value()->getEdges()->size()) / (numVertices - 1);
+
+		float nodeDegree = ((float) i.value()->getEdges()->size()) / (float)(numVertices - 1);
 		i.value()->setNodeDegree(nodeDegree);
 
 		// get all shortest paths
@@ -36,7 +41,7 @@ void GraphMetrics::computeGraphMetrics(Data::Graph * activeGraph)
 		paths[idx] = new QVector<QLinkedList<osg::ref_ptr<Data::Node> > > [numVertices];
 
 		// compute node closeness
-		int pathSum = 0;
+		qlonglong pathSum = 0;
 		while(i2.hasNext())
 		{
 			i2.next();
@@ -46,7 +51,7 @@ void GraphMetrics::computeGraphMetrics(Data::Graph * activeGraph)
 		if (numVertices == 1)
 			numVertices = 2;
 
-		i.value()->setNodeCloseness(((float) pathSum) / (numVertices - 1));
+		i.value()->setNodeCloseness(((float) pathSum) / (float)(numVertices - 1));
 
 		// check for regexp match
 		if (re.exactMatch(i.value()->getName()))
@@ -100,7 +105,7 @@ void GraphMetrics::computeGraphMetrics(Data::Graph * activeGraph)
 			}
 		}
 
-		i.value()->setNodeBetweeness(sum / ((float) (numVertices - 1) * (numVertices - 2)));
+		i.value()->setNodeBetweeness(sum / ((float) (numVertices - 1) * (float)(numVertices - 2)));
 
 		float base = i.value()->isNodeMatched() ? regexpWeigth : 0;
