+++ src/QOSG/SelectLayoutWindow.cpp
@@ -1,5 +1,10 @@
 #include "QOSG/SelectLayoutWindow.h"
 
+#include "Model/GraphLayoutDAO.h"
+#include "Model/DB.h"
+#include "Model/GraphDAO.h"
+#include "Manager/Manager.h"
+
 using namespace QOSG;
 
 qlonglong graphID;
@@ -61,7 +66,7 @@ void SelectLayoutWindow::createLayoutTable()
 	layoutsCount = layouts.count();
 
 	numberOfLayouts->setText(tr("%1 layout(s) found").arg(layoutsCount));
-	layoutsTable->setRowCount(layoutsCount);
+	layoutsTable->setRowCount((int)layoutsCount);
 
 	for(iterLayout = layouts.begin(), row=0; iterLayout != layouts.end(); ++iterLayout, row++)
 	{
@@ -71,8 +76,8 @@ void SelectLayoutWindow::createLayoutTable()
 
 		QTableWidgetItem *itemID = new QTableWidgetItem(tr("%1").arg(layoutID));
 		QTableWidgetItem *itemName = new QTableWidgetItem(name);
-		layoutsTable->setItem(row, 0, itemID);
-		layoutsTable->setItem(row, 1, itemName);
+		layoutsTable->setItem((int)row, 0, itemID);
+		layoutsTable->setItem((int)row, 1, itemName);
 	}
 	if(layoutsTable->rowCount() > 0)
 		layoutsTable->selectRow(0);
